refactor(ai): move ModelAuthoredSchema under task/base
Its test moves with it, since base modules keep theirs in base/__tests__.
