+++ packages/ai/src/task/base/ModelAuthoredSchema.ts
