+++ packages/ai/src/task/base/__tests__/ModelAuthoredSchema.test.ts
