+++ packages/ai/src/task/index.ts
@@ -16,6 +16,7 @@ export * from "./base/AiTask";
 export * from "./base/AiTaskSchemas";
 export * from "./base/chatTurn";
 export * from "./base/CheckpointPorts";
+export * from "./base/ModelAuthoredSchema";
 export * from "./base/responseFormat";
 export * from "./base/runWithIterable";
 export * from "./base/StreamingAiTask";
@@ -46,7 +47,6 @@ export * from "./KbReindexTask";
 export * from "./KbSearchTask";
 export * from "./KbToDocumentsTask";
 export * from "./MessageConversion";
-export * from "./ModelAuthoredSchema";
 export * from "./ModelDownloadRemoveTask";
 export * from "./ModelDownloadTask";
 export * from "./ModelInfoTask";
